Allow strings longer than 2 characters for the HTTP probe
Previously only content of lenght 2 or less characters was allowed for
the HTTP probe. The length was hardcoded and even if the strings from
the configuration and from the probe matched, the result was failure.

This could be reproduced e.g. with using the following line in the
configuration:

url: "http://httpstat.us/200 200 OK"

Signed-off-by: Tomas Hozza <thozza@redhat.com>